strutil,cmd/snap-update-ns: move PathIterator to strutil
The path iterator is a generic and useful utility and I actually need to
use it in the apparmor backend now. I was considering osutil but since
it's not really touching the file system in any way I opted for generic
strutil instead.
