+++ cmd/snap-update-ns/utils_test.go
@@ -1002,138 +1002,3 @@ func (s *realSystemSuite) TestSecureOpenPathSymlinkedParent(c *C) {
 	c.Check(fd, Equals, -1)
 	c.Check(err, ErrorMatches, "not a directory")
 }
-
-func (s *realSystemSuite) TestPathIteratorEmpty(c *C) {
-	iter, err := update.NewPathIterator("")
-	c.Assert(err, ErrorMatches, `cannot iterate over unclean path ""`)
-	c.Assert(iter, IsNil)
-}
-
-func (s *realSystemSuite) TestPathIteratorFilename(c *C) {
-	iter, err := update.NewPathIterator("foo")
-	c.Assert(err, IsNil)
-	c.Assert(iter.Path(), Equals, "foo")
-
-	c.Assert(iter.Next(), Equals, true)
-	c.Assert(iter.CurrentBase(), Equals, "")
-	c.Assert(iter.CurrentPath(), Equals, "foo")
-	c.Assert(iter.CurrentName(), Equals, "foo")
-	c.Assert(iter.CurrentCleanName(), Equals, "foo")
-
-	c.Assert(iter.Next(), Equals, false)
-}
-
-func (s *realSystemSuite) TestPathIteratorRelative(c *C) {
-	iter, err := update.NewPathIterator("foo/bar")
-	c.Assert(err, IsNil)
-	c.Assert(iter.Path(), Equals, "foo/bar")
-
-	c.Assert(iter.Next(), Equals, true)
-	c.Assert(iter.CurrentBase(), Equals, "")
-	c.Assert(iter.CurrentPath(), Equals, "foo/")
-	c.Assert(iter.CurrentName(), Equals, "foo/")
-	c.Assert(iter.CurrentCleanName(), Equals, "foo")
-
-	c.Assert(iter.Next(), Equals, true)
-	c.Assert(iter.CurrentBase(), Equals, "foo/")
-	c.Assert(iter.CurrentPath(), Equals, "foo/bar")
-	c.Assert(iter.CurrentName(), Equals, "bar")
-	c.Assert(iter.CurrentCleanName(), Equals, "bar")
-
-	c.Assert(iter.Next(), Equals, false)
-}
-
-func (s *realSystemSuite) TestPathIteratorAbsoluteClean(c *C) {
-	iter, err := update.NewPathIterator("/foo/bar")
-	c.Assert(err, IsNil)
-	c.Assert(iter.Path(), Equals, "/foo/bar")
-
-	c.Assert(iter.Next(), Equals, true)
-	c.Assert(iter.CurrentBase(), Equals, "")
-	c.Assert(iter.CurrentPath(), Equals, "/")
-	c.Assert(iter.CurrentName(), Equals, "/")
-	c.Assert(iter.CurrentCleanName(), Equals, "")
-
-	c.Assert(iter.Next(), Equals, true)
-	c.Assert(iter.CurrentBase(), Equals, "/")
-	c.Assert(iter.CurrentPath(), Equals, "/foo/")
-	c.Assert(iter.CurrentName(), Equals, "foo/")
-	c.Assert(iter.CurrentCleanName(), Equals, "foo")
-
-	c.Assert(iter.Next(), Equals, true)
-	c.Assert(iter.CurrentBase(), Equals, "/foo/")
-	c.Assert(iter.CurrentPath(), Equals, "/foo/bar")
-	c.Assert(iter.CurrentName(), Equals, "bar")
-	c.Assert(iter.CurrentCleanName(), Equals, "bar")
-
-	c.Assert(iter.Next(), Equals, false)
-}
-
-func (s *realSystemSuite) TestPathIteratorAbsoluteUnclean(c *C) {
-	iter, err := update.NewPathIterator("/foo/bar/")
-	c.Assert(err, ErrorMatches, `cannot iterate over unclean path "/foo/bar/"`)
-	c.Assert(iter, IsNil)
-}
-
-func (s *realSystemSuite) TestPathIteratorRootDir(c *C) {
-	iter, err := update.NewPathIterator("/")
-	c.Assert(err, IsNil)
-	c.Assert(iter.Path(), Equals, "/")
-
-	c.Assert(iter.Next(), Equals, true)
-	c.Assert(iter.CurrentBase(), Equals, "")
-	c.Assert(iter.CurrentPath(), Equals, "/")
-	c.Assert(iter.CurrentName(), Equals, "/")
-	c.Assert(iter.CurrentCleanName(), Equals, "")
-
-	c.Assert(iter.Next(), Equals, false)
-}
-
-func (s *realSystemSuite) TestPathIteratorUncleanPath(c *C) {
-	iter, err := update.NewPathIterator("///some/../junk")
-	c.Assert(err, ErrorMatches, `cannot iterate over unclean path ".*"`)
-	c.Assert(iter, IsNil)
-}
-
-func (s *realSystemSuite) TestPathIteratorUnicode(c *C) {
-	iter, err := update.NewPathIterator("/zażółć/gęślą/jaźń")
-	c.Assert(err, IsNil)
-	c.Assert(iter.Path(), Equals, "/zażółć/gęślą/jaźń")
-
-	c.Assert(iter.Next(), Equals, true)
-	c.Assert(iter.CurrentBase(), Equals, "")
-	c.Assert(iter.CurrentPath(), Equals, "/")
-	c.Assert(iter.CurrentName(), Equals, "/")
-	c.Assert(iter.CurrentCleanName(), Equals, "")
-
-	c.Assert(iter.Next(), Equals, true)
-	c.Assert(iter.CurrentBase(), Equals, "/")
-	c.Assert(iter.CurrentPath(), Equals, "/zażółć/")
-	c.Assert(iter.CurrentName(), Equals, "zażółć/")
-	c.Assert(iter.CurrentCleanName(), Equals, "zażółć")
-
-	c.Assert(iter.Next(), Equals, true)
-	c.Assert(iter.CurrentBase(), Equals, "/zażółć/")
-	c.Assert(iter.CurrentPath(), Equals, "/zażółć/gęślą/")
-	c.Assert(iter.CurrentName(), Equals, "gęślą/")
-	c.Assert(iter.CurrentCleanName(), Equals, "gęślą")
-
-	c.Assert(iter.Next(), Equals, true)
-	c.Assert(iter.CurrentBase(), Equals, "/zażółć/gęślą/")
-	c.Assert(iter.CurrentPath(), Equals, "/zażółć/gęślą/jaźń")
-	c.Assert(iter.CurrentName(), Equals, "jaźń")
-	c.Assert(iter.CurrentCleanName(), Equals, "jaźń")
-
-	c.Assert(iter.Next(), Equals, false)
-}
-
-func (s *realSystemSuite) TestPathIteratorExample(c *C) {
-	iter, err := update.NewPathIterator("/some/path/there")
-	c.Assert(err, IsNil)
-	for iter.Next() {
-		_ = iter.CurrentBase()
-		_ = iter.CurrentPath()
-		_ = iter.CurrentName()
-		_ = iter.CurrentCleanName()
-	}
-}
